Fix user agent (read extended description)
I did not try to rebrand the code. This is the GameTDB user agent, which user agents are required to download from it. I wanted to differentiate it to make it not seem weird that 2 scripts on different IPs had the exact user agent. WiiLink's name is mentioned on the main.go which I did not actually change, but I'm reverting this part because of the drama WiiLink started over this.